refactor: extract app router into separate component
